// src/redux/movies/types.ts
enum MoviesTypes {
  GET_MOVIES = 'GET_MOVIES',
  GET_MOVIES_SUCCESS = 'GET_MOVIES_SUCCESS',
  GET_MOVIES_ERROR = 'GET_MOVIES_ERROR',

  //GET_GENRES
  GET_GENRES = 'GET_GENRES',
  GET_GENRES_SUCCESS = 'GET_GENRES_SUCCESS',
  GET_GENRES_ERROR = 'GET_GENRES_ERROR',

  // GET_NOW_PLAYING
  GET_SECONDARY_SLIDER_DATA = 'GET_SECONDARY_SLIDER_DATA',
  GET_SECONDARY_SLIDER_DATA_SUCCESS = 'GET_SECONDARY_SLIDER_DATA_SUCCESS',
  GET_SECONDARY_SLIDER_DATA_ERROR = 'GET_SECONDARY_SLIDER_DATA_ERROR',
}
export default MoviesTypes;